add checker for scala sbt auto recompile
sbt is capable of watching all the scala files in the project and
automatically recompiling things as they change. Obviously, since it's
incremental compilation, rather than batch compilation, it's *much*
faster than a batch compilation. When sbt performs this compilation, it
prints errors out to a temporary file in the project's target directory.

This patch adds a syntax_checker which (when executed) reads errors from
this temporary file.